Move several functions to helix-core
