+++ helix-core/src/lib.rs
@@ -10,6 +10,7 @@ pub mod line_ending;
 pub mod macros;
 pub mod match_brackets;
 pub mod movement;
+pub mod numbers;
 pub mod object;
 pub mod path;
 mod position;
